Establish pyramidal data source abstraction
